requirements.txt
Pillow==2.7.0
wheel==0.24.0
